Added gender identity as a described class
Some people dispute that discrimination against people who are transgender is not discrimination based on gender, because they disagree with the definition of "gender" as defined by other contexts. Specifically calling out "gender identity" offers explicit protection for transgender people, since it tells possible offenders that a different definition of "gender" is not a loophole.